requirements.txt
alembic==1.13.0
attrs==23.1.0
black==23.11.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
clickclick==20.10.2
envparse==0.2.0
Flask==2.2.0
Flask-Login==0.6.3
Flask-Migrate==2.7.0
Flask-SQLAlchemy==3.0.3
flask-swagger-ui==4.11.1
greenlet==3.0.1
idna==3.6
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
Mako==1.3.0
MarkupSafe==2.1.3
marshmallow==3.15.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==4.0.0
psycopg2-binary==2.9.9
PyYAML==6.0.1
referencing==0.31.1
requests==2.31.0
rpds-py==0.13.2
SQLAlchemy==2.0.23
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.1.0
Werkzeug==2.2.0
